+++ CHANGELOG.md
@@ -24,6 +24,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- [#51](https://github.com/XenitAB/spegel/pull/51) Filter tracked images to only included mirrored registries.
+
 ### Security
 
 ## v0.0.4